# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Argus Personal AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  completed?: boolean;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  dueDate?: string | null;
}