# lib/api-zod/src/generated/types/updateWorkflowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Argus Personal AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWorkflowBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  trigger?: string;
  steps?: string;
  enabled?: boolean;
}