# lib/api-zod/src/generated/types/workflow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Argus Personal AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Workflow {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  trigger: string;
  steps: string;
  enabled: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  lastRunStatus?: string | null;
  createdAt: string;
}